Add custom Note type
The Note type provided by DBus.Notify is more complicated than we
want to support, at least for now. With an eye to the future, we
may want to support different notification systems (e.g. OSX), so
it behooves us to create a custom Note type that the majority of
our application will use (including the config file), and transform
it to the DBus's Note at the last moment.

There are a few additional changes:

- Make the TOML parsing stricter. Previously unrecognized values did not
  throw errors. This is misleading, so we want to enforce strict parsing.
- Add Show instances to more types for debugging, along with more logging.
- Clean up some code via Optics.