Swap Generic Optics for TH Labels
The current optics was based on Optics-0.4's generic functionality,
i.e., given a type that is an instance of Generic, we can access its
fields via `aRecord ^. gfield @"aField"`. This is nice because we do
not have to write any optics ourselves or generate them with
TemplateHaskell. Unfortunately, this becomes a problem when we cannot
derive Generic for a type because it includes a third-party field that
does not have a Generic instance.

If the Generic instance could be provided then we can either use
-XOverlappingInstances or wrap it in a newtype. Both of these
workarounds have tradeoffs, but they do not help with the type cannot be
given a generic instance, which is situation we found ourselves in.

Fortunately, we can use Optic's label functionality instead, generating
the instances via TH. This has the nice bonus of considerably improving
ergonomics, as we can now simply write `aRecord ^. #aField`.