Remove unnecessary field disambiguation
Previously, a prefix was added to Env's log fields to differeniate one
of the fields (logEnv) from a Katip typeclass method. This is no longer
as necessary as we can get around this using OverloadedLabels. We may
still run into annoyances if both are in scope and we're constructing
an Env using the field name, but that's a rare case, not worth making
the field names worse.