Each services now has its own timeout
Allowing each service to have its own timeout (a desirement for a
long time) has had a significant impact on the code base.

First, the most straightforward way to do this is run each service
in its own thread, thus we need concurrency. This part is actually
really straightforward, except we need to control access to shared
services i.e. logging and sending notifications. For these, we
introduce queues that can handle the gory details themselves;
all we have to do is modify our existing functions to instead
send messages to the queue, then we have dedicated threads that
continuously poll the queues and send the data where it needs
to go.

Beyond that, we also need to add poll-interval config to the
toml for each service.

Since we're using unliftio for the async logic, we can remove
safe-exceptions as unliftio provides exception logic.

We also add strict annotations to most data.