Use STM retry logic when reading from queue
Previously, we were using reads that can fail (i.e. do not block). The
naive infinite loop had heavy cpu usage as it was constantly running.
To combat this, we introduced a poll-interval, that would sleep for the
given interval an check the queue once it woke up. This worked okay, in
that it helped temper the cpu usage, though it had some limitations:

* It wasn't clear what an ideal interval would be (anything > 1 second)
  seemed fine.
* Complicated API.
* We no longer necessarily had prompt logging / notification sending.
  That is, we could end up in a situation where a notification is
  triggered yet the thread reading the queue was still asleep.

We briefly experimented with using STM's TSem semaphore to include a
signal with queue writes. This would completely remove the need for
the interval, meaning the API would be simpler and we would no longer
have the promptness issues.

Happily, TBQueue already has this functionality; the default read
blocks until a value exists. Thus we simply replace our maybe'd
STM.tryReadTBQueue with STM.readTBQueue, which blocks at a low cpu
usage.