Swap tomland for toml-reader
Toml-reader has a far simpler API than tomland as it handles decoding
only, not encoding (we only need the former).

Furthermore, we can easily improve the decoding to better match the real
types (e.g. parse our timeout as toml string or integer, not just
string). This is probably possible with tomland, but it's less obvious.

Finally, we can also immediately upgrade to ghc 9.2 as toml-reader is
compatible, unlike tomland.