Change updatePrevTrigger logic
Previously, we were always storing the query result. But this is
technically unnecessary, as we only need to store a result that leads
to raising an alert.

There is no semantic difference between storing an 'uninteresting'
result (wrt raising an alert) and storing Nothing. Both are going to
'reset' the repeat blocking behavior.

Moreover, this change will more easily allow us to generalize the
repeat blocking behavior to allow battery percentage ranges.